bot: A2 mechanical module split of PlaybookBotExecution (zero behaviour change)
Split src/core/execution/playbook/PlaybookBotExecution.ts (1,457 lines) into:
- Params.ts      PlaybookParams + DEFAULT_PLAYBOOK (re-exported from PlaybookBotExecution.ts)
- Context.ts     BotContext interface (mg, me, p, sit, random, send, boat, log)
- Situation.ts   Situation interface + SituationQueries (neighbours, cap, density, outgoingTo,
                 annexable, landmassSize/landmassTiles, acrossWater, postFacing)
- Military.ts    expand, harvestBots, counterAttack, collapsed, fight, manageRetreats,
                 earlyBoat, sendBoat, huntBotsByBoat, seaInvasion, seaExpansion,
                 maybeBomb, clearOfFriends, maybeMIRV, watchSplit, noteSent/reachable
- Economy.ts     build, buildRail, railInfillTile, railFactorySpot, cityUnitCap, rank,
                 tryBuild, sampleTerritory, tileNear, interiorTile, oceanShoreTile,
                 portTile, defensePostTile
- Diplomacy.ts   acceptAlliances, isPrey, requestAlliances, manageExpiries,
                 manageEmbargoes, onAllianceEnded
PlaybookBotExecution.ts keeps init/tick/isActive, readSituation, send, boat, events,
the rule table, and the static pickSpawn/basin/inland/lastSpawnDiag. External imports
(GameRunner, lab tests, PlaybookBotHook test) are unchanged.

Code was moved, not improved: no thresholds, log strings, rule order or variable names
changed. `if (this.log.length < 2000) this.log.push(x)` became `ctx.log(x)` (same cap).

Shared private state moves (owner = module that writes it; getter where read elsewhere):
- currentTarget  -> Military (written by fight/maybeBomb); getter read by Diplomacy, Economy
- plannedTarget  -> Diplomacy (manageExpiries); getter read by Military via a () => accessor
- counters       -> Military (counterAttack, manageRetreats, fight) — internal
- lastWarTick    -> Military (fight) — internal
- postFailed     -> Economy (build); map getter, Diplomacy.onAllianceEnded deletes from it
- embargoedAt    -> Military (fight); map getter read by Diplomacy.manageEmbargoes
- bombOutOfRange -> Military (maybeBomb); getter + setter, Economy.build reads and resets it
- bombs          -> Military (maybeBomb, maybeMIRV); PlaybookBotExecution.bombs is now a getter
- waves, sentAt, blacklist, lastCounter, bombed, boatedAt, lastSeaTick, lastMirvTick,
  lastBombTick, lastInvasionTick, attackStart, history, splitOwner/splitTile/splitSince
                 -> Military — internal
- rail, pendingAnchor(Tick), pendingFactory(Tick), railDiag, firstPortTick, lastSamTick,
  lastWarshipTick -> Economy — internal
- annexCache     -> SituationQueries (annexable is read by Military and Diplomacy)
- boatSent, landmassChecked, onSmallLandmass, lastMode, prevAllies, prevIncoming, kills,
  log            -> stay in PlaybookBotExecution

Deviations from a pure move (all behaviour-neutral):
- Military.maybeBomb reads plannedTarget through a function, so the null check uses a
  local `const plannedTarget = this.plannedTarget()` (TS narrowing).
- readSituation stays in PlaybookBotExecution (per the worker brief) because it calls
  Military.collapsed; Situation.ts holds the interface and the queries.

Proof: lab game (africa, Medium) at MIN=1, 3 and 8 before/after the split — transcripts
byte-identical except botMs/gameMs. tests/PlaybookBotHook.test.ts green, tsc clean,
lint clean apart from a pre-existing error in tests/lab/growth.lab.test.ts.
